backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
pandas==2.2.2
openpyxl==3.1.2
slowapi==0.1.9
python-dotenv==1.0.1
